Feat: Scheduler class 생성
